v1.0
this is the first fully working version. it allows the playback of audio along with making the sections from the time stamps in that audio. there's not really any error handling or checks against user error, that may or may not come later. slider still needs to change the song time.