disabled buttons after section create
disabled buttons after section create. also fixed bug that made it so you couldn't split another song after the first one(would crash). i'm going to keep the debug folder in, reason being that if someone wanted to actually open the vs project up and debug it, they would need the ffmpeg executables in there. instead of them having to figure it out i'll just keep the folder in source control so that it's not a problem.